aliases: mkpw: random password; approx. brute-force chance: 1/(16**128)
Analysis of brute-force complexity is based on 128 characters of output
(using sha512sum, so 512 bytes in 4 bit increments [hexadecimal] is 128
total characters of output), and 16 choices for each character,
effectively randomly distributed based on the input being bytes read out
of /dev/urandom and the input size being the same size as the output
size. In other words: 128 "slots" of 16 choices each; independently
sampled; which is 16*16*...*16 or 16**128. The odds of guessing the
exact right sequence is around 1 in 16**128, assuming /dev/urandom is a
fairly reliable source of randomness. (Which it probably technically
isn't.)

Analysis is likely flawed. I am not a cryptographer. (erm, ... "IANAC"?
something like that.)